+++ crypto-ffi/src/bundles/commit.rs
@@ -23,8 +23,8 @@ impl TryFrom<MlsCommitBundle> for CommitBundle {
     fn try_from(msg: MlsCommitBundle) -> Result<Self, Self::Error> {
         let encrypted_message = msg.encrypted_message.clone();
         let (welcome, commit, group_info) = msg.to_bytes_triple()?;
-        let welcome = welcome.map(Welcome::from).map(Arc::new);
-        let group_info = group_info.into();
+        let welcome = welcome.map(Into::into).map(Arc::new);
+        let group_info = group_info.try_into()?;
         Ok(Self {
             welcome,
             commit,