+++ crypto/src/mls/conversation/mod.rs
@@ -58,10 +58,26 @@ pub use self::{
 use super::credential::Credential;
 use crate::{
     Ciphersuite, ClientId, ClientIdRef, CredentialRef, CredentialType, E2eiConversationState, LeafError, MlsError,
-    RecursiveError, UserId, WireIdentity,
+    RecursiveError, UserId, WireIdentity, bytes_wrapper,
     mls::{HasSessionAndCrypto, Session, credential::ext::CredentialExt as _},
     mls_provider::MlsCryptoProvider,
 };
+
+bytes_wrapper!(
+    /// A secret key derived from the group secret.
+    ///
+    /// This is intended to be used for AVS.
+    #[derive(Clone)]
+    SecretKey
+);
+
+bytes_wrapper!(
+    /// The raw public key of an external sender.
+    ///
+    /// This can be used to initialize a subconversation.
+    #[derive(Clone)]
+    ExternalSenderKey
+);
 
 /// The base layer for [Conversation].
 /// The trait is only exposed internally.
@@ -134,7 +150,7 @@ pub trait Conversation<'a>: ConversationWithMls<'a> {
     ///
     /// # Errors
     /// OpenMls secret generation error
-    async fn export_secret_key(&'a self, key_length: usize) -> Result<Vec<u8>> {
+    async fn export_secret_key(&'a self, key_length: usize) -> Result<SecretKey> {
         const EXPORTER_LABEL: &str = "exporter";
         const EXPORTER_CONTEXT: &[u8] = &[];
         let backend = self.crypto_provider().await?;
@@ -142,6 +158,7 @@ pub trait Conversation<'a>: ConversationWithMls<'a> {
         inner
             .group()
             .export_secret(&backend, EXPORTER_LABEL, EXPORTER_CONTEXT, key_length)
+            .map(Into::into)
             .map_err(MlsError::wrap("exporting secret key"))
             .map_err(Into::into)
     }
@@ -161,7 +178,7 @@ pub trait Conversation<'a>: ConversationWithMls<'a> {
 
     /// Returns the raw public key of the single external sender present in this group.
     /// This should be used to initialize a subconversation
-    async fn get_external_sender(&'a self) -> Result<Vec<u8>> {
+    async fn get_external_sender(&'a self) -> Result<ExternalSenderKey> {
         let inner = self.conversation().await;
         let ext_senders = inner
             .group()
@@ -169,7 +186,7 @@ pub trait Conversation<'a>: ConversationWithMls<'a> {
             .external_senders()
             .ok_or(Error::MissingExternalSenderExtension)?;
         let ext_sender = ext_senders.first().ok_or(Error::MissingExternalSenderExtension)?;
-        let ext_sender_public_key = ext_sender.signature_key().as_slice().to_vec();
+        let ext_sender_public_key = ext_sender.signature_key().as_slice().to_vec().into();
         Ok(ext_sender_public_key)
     }